+++ benchmark_test.go
@@ -404,7 +404,9 @@ func benchmarkTransport(b *testing.B, sz int) {
 	addr2 := "localhost:43568"
 	nhc1 := config.NodeHostConfig{
 		RaftAddress: addr1,
-		FS:          vfs.DefaultFS,
+		Expert: config.ExpertConfig{
+			FS: vfs.DefaultFS,
+		},
 	}
 	env1, err := server.NewEnv(nhc1, vfs.DefaultFS)
 	if err != nil {
@@ -412,7 +414,9 @@ func benchmarkTransport(b *testing.B, sz int) {
 	}
 	nhc2 := config.NodeHostConfig{
 		RaftAddress: addr2,
-		FS:          vfs.DefaultFS,
+		Expert: config.ExpertConfig{
+			FS: vfs.DefaultFS,
+		},
 	}
 	env2, err := server.NewEnv(nhc2, vfs.DefaultFS)
 	if err != nil {