+++ snapshotter_test.go
@@ -47,7 +47,6 @@ func getNewTestDB(dir string, lldir string, fs vfs.IFS) raftio.ILogDB {
 		panic(err)
 	}
 	cfg := config.NodeHostConfig{
-		LogDB:  config.GetDefaultLogDBConfig(),
 		Expert: config.GetDefaultExpertConfig(),
 	}
 	db, err := logdb.NewDefaultLogDB(cfg, nil, []string{d}, []string{lld}, fs)
@@ -71,7 +70,8 @@ func getTestSnapshotter(ldb raftio.ILogDB, fs vfs.IFS) *snapshotter {
 	f := func(cid uint64, nid uint64) string {
 		return fp
 	}
-	return newSnapshotter(1, 1, config.NodeHostConfig{FS: fs}, f, ldb, fs)
+	return newSnapshotter(1,
+		1, config.NodeHostConfig{Expert: config.ExpertConfig{FS: fs}}, f, ldb, fs)
 }
 
 func runSnapshotterTest(t *testing.T,