+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Note that Pebble provides bidirectional compatibility with RocksDB. Existing Dra
 - The signature of config.LogDBFactoryFunc has been changed. Your application is not affected unless it uses a custom LogDB implementation.
 - Due to lack of user interests, C++ binding is no longer supported.
 - LevelDB based LogDB is no longer supported.
+- NodeHostConfig's FS and SystemTickerPrecision fields have been moved into NodeHostConfig.Expert.
 
 ## v3.2 (2020-03-05)